Creating unit testing while keeping e2e testing
