+++ jest.config.js
@@ -21,7 +21,7 @@ const config = {
 
   // An array of glob patterns indicating a set of files for which coverage information should be collected
   collectCoverageFrom: [
-    "../dist/**" 
+    "./dist/**" 
   ],
 
   // The directory where Jest should output its coverage files
