+++ package.json
@@ -32,8 +32,11 @@
   ],
   "scripts": {
     "start": "npm run build && node dist/index.js",
-    "test": "npm run build && npm test --prefix ./e2e/",
-    "test:update": "npm run build && npm run test:update --prefix ./e2e/",
+    "pree2e": "npm run build",
+    "e2e": "node --experimental-vm-modules node_modules/jest/bin/jest.js --detectOpenHandles --testPathPattern /src/e2e/",
+    "e2e:update": "npm run e2e -- --updateSnapshot",
+    "pretest": "npm run build",
+    "test": "node node_modules/jest/bin/jest.js --detectOpenHandles --testPathIgnorePatterns /src/e2e/ --testPathPattern /src/",
     "pack": "npm run build && npm pack --pack-destination dist",
     "build": "npx eslint ./src && tsc --project tsconfig.json",
     "postbuild": "npm run format-check",
@@ -49,11 +52,15 @@
     "@types/minimist": "^1.2.5",
     "@types/node": "^20.11.16",
     "@types/yargs": "^17.0.32",
+    "@types/jest": "^29.5.14",
     "beachball": "^2.47.1",
     "eslint": "^8.57.0",
     "prettier": "^3.2.5",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.9.0"
+    "typescript-eslint": "^7.9.0",
+    "execa": "^9.5.2",
+    "ts-jest": "^29.2.5",
+    "jest": "^29.7.0"
   },
   "dependencies": {
     "chalk": "^4.1.2",
